# MyApi/api-spec.yaml
openapi: 3.0.4
info:
  title: My API (External)
  description: External API - Version 1.0
  version: '1.0'
paths:
...
  /v1/device/private:
    get:
      tags:
        - Device
      summary: Get private devices.
      description: This endpoint is protected and requires a valid token with the "ApiScope" policy.
      responses:
        '200':
          description: Returns the list of devices and user details.
          content:
            application/json:
              schema: { }
        '401':
          description: Unauthorized if the token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/device/internal/management:
    get:
      tags:
        - Device
      summary: Get internal device management data (Internal API only).
      description: "This endpoint provides sensitive internal device management information\r\nand is only accessible to clients with internal scope."
      responses:
        '200':
          description: Returns internal device data.
        '403':
          description: Forbidden - Client doesn't have internal access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/device/external/partner-data:
    get:
      tags:
        - Device
      summary: Get external partner device data (External API only).
      description: "This endpoint provides device information formatted for external partners\r\nand is only accessible to clients with external scope."
      responses:
        '200':
          description: Returns external device data.
        '403':
          description: Forbidden - Client doesn't have external access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/v1/device/internal/{deviceId}/config':
    put:
      tags:
        - Device
      summary: Update device configuration (Internal write access required).
      description: "This endpoint allows updating device configuration and requires both\r\ninternal scope and write permissions."
      parameters:
        - name: deviceId
          in: path
          description: The ID of the device to update.
          required: true
          schema:
            type: string
      requestBody:
        description: The new device configuration.
        content:
          application/json:
            schema: { }
          text/json:
            schema: { }
          application/*+json:
            schema: { }
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /v1/device/with-params:
    post:
      tags:
        - Device
      summary: Create a device with additional parameters.
      description: This endpoint allows you to create a device by providing additional parameters such as location and status.
      parameters:
        - name: location
          in: query
          description: The location for the device.
          schema:
            type: string
        - name: status
          in: query
          description: 'The status of the device (0=Offline, 1=Online).'
          schema:
            type: integer
            format: int32
            default: 0
      requestBody:
        description: The request body containing the Device details.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Device'
          text/json:
            schema:
              $ref: '#/components/schemas/Device'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Device'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/v1/device/{id}/{date}':
    delete:
      tags:
        - Device
      summary: Delete a device by ID and date.
      description: "This endpoint deletes a device based on the provided ID and date.\r\nThe date must be in the format `yyyy-MM-dd`.\r\nExample: `/devices/1/2025-04-30`."
      parameters:
        - name: id
          in: path
          description: The ID of the device to delete.
          required: true
          schema:
            type: integer
            format: int32
            default: 2
        - name: date
          in: path
          description: The date of the device registration to delete.
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '204':
          description: No content - The device was successfully deleted.
        '400':
          description: Bad request - The date format is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not found - The device with the specified ID and date does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    Device:
      type: object
      properties:
        name:
          type: string
          nullable: true
        deviceType:
          $ref: '#/components/schemas/DeviceType'
        dimension:
          $ref: '#/components/schemas/Dimension'
      additionalProperties: false
      description: Represents a device.
    DeviceType:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the device type
          format: int32
        name:
          type: string
          description: The name of the device type
          nullable: true
      additionalProperties: false
      description: Represents a device type.
    Dimension:
      type: object
      properties:
        width:
          type: number
          description: The width of the device in centimeters
          format: double
        height:
          type: number
          description: The height of the device in centimeters
          format: double
        depth:
          type: number
          description: The depth of the device in centimeters
          format: double
        volume:
          type: number
          description: Gets the volume of the device in cubic centimeters.
          format: double
          readOnly: true
      additionalProperties: false
      description: Represents the physical dimensions of a device.
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://localhost:5001/connect/token
          scopes:
            devices.read: Read access to the Devices Management API
            devices.write: Write access to the Devices Management API
            devices.internal: Internal API Access
            devices.external: External API Access
security:
  - oauth2:
      - devices.read
      - devices.write
      - devices.internal
      - devices.external